support read from specific timestamp
