+++ hstream/hstream.cabal
@@ -80,6 +80,7 @@ library
     HStream.Server.Core.Stream
     HStream.Server.Core.Subscription
     HStream.Server.Core.View
+    HStream.Server.Core.ShardReader
     HStream.Server.Handler.Admin
     HStream.Server.Handler.Cluster
     HStream.Server.Handler.Connector
@@ -89,6 +90,7 @@ library
     HStream.Server.Handler.Stream
     HStream.Server.Handler.Subscription
     HStream.Server.Handler.View
+    HStream.Server.Handler.ShardReader
     HStream.Server.MetaData.Exception
     HStream.Server.MetaData.Types
     HStream.Server.MetaData.Utils
