+++ hstream/src/HStream/Server/Handler/Cluster.hs
@@ -60,7 +60,6 @@ lookupSubscriptionHandler
 lookupSubscriptionHandler sc (ServerNormalRequest _meta req) =
   defaultExceptionHandle $ returnResp =<< C.lookupSubscription sc req
 
-{-# DEPRECATED lookupShardReaderHandler "Use lookupResourceHandler instead" #-}
 lookupShardReaderHandler
   :: ServerContext
   -> ServerRequest 'Normal LookupShardReaderRequest LookupShardReaderResponse
@@ -85,7 +84,7 @@ handleLookupSubscription
   -> G.UnaryHandler LookupSubscriptionRequest LookupSubscriptionResponse
 handleLookupSubscription sc _ req = catchDefaultEx $ C.lookupSubscription sc req
 
-{-# DEPRECATED handleLookupShardReader "Use handleLookupResource instead" #-}
+-- handleLookupShardReader should only be called before readShardStream request
 handleLookupShardReader
   :: ServerContext
   -> G.UnaryHandler LookupShardReaderRequest LookupShardReaderResponse
