+++ hstream/src/HStream/Server/MetaData/Value.hs
@@ -10,7 +10,7 @@ import           HStream.MetaStore.Types       (FHandle, HasPath (myRootPath),
 import           HStream.Server.MetaData.Types
 import           HStream.Server.Types          (SubscriptionWrap (..))
 import qualified HStream.ThirdParty.Protobuf   as Proto
-import           HStream.Utils                 (TaskStatus (..), textToCBytes)
+import           HStream.Utils                 (textToCBytes)
 
 paths :: [CBytes]
 paths = [ textToCBytes rootPath
@@ -18,7 +18,7 @@ paths = [ textToCBytes rootPath
         , textToCBytes $ myRootPath @TaskIdMeta       @ZHandle
         , textToCBytes $ myRootPath @TaskMeta         @ZHandle
         , textToCBytes $ myRootPath @TaskKvMeta       @ZHandle
-        , textToCBytes $ myRootPath @ShardReader      @ZHandle
+        , textToCBytes $ myRootPath @ShardReaderMeta      @ZHandle
         , textToCBytes $ myRootPath @QueryInfo        @ZHandle
         , textToCBytes $ myRootPath @QueryStatus      @ZHandle
         , textToCBytes $ myRootPath @ViewInfo         @ZHandle
@@ -35,7 +35,7 @@ tables = [
   , myRootPath @QueryInfo        @RHandle
   , myRootPath @QueryStatus      @RHandle
   , myRootPath @ViewInfo         @RHandle
-  , myRootPath @ShardReader      @RHandle
+  , myRootPath @ShardReaderMeta      @RHandle
   , myRootPath @SubscriptionWrap @RHandle
   , myRootPath @Proto.Timestamp  @RHandle
   , myRootPath @TaskAllocation   @RHandle
@@ -49,7 +49,7 @@ fileTables = [
   , myRootPath @QueryInfo        @FHandle
   , myRootPath @QueryStatus      @FHandle
   , myRootPath @ViewInfo         @FHandle
-  , myRootPath @ShardReader      @FHandle
+  , myRootPath @ShardReaderMeta      @FHandle
   , myRootPath @SubscriptionWrap @FHandle
   , myRootPath @Proto.Timestamp  @FHandle
   , myRootPath @TaskAllocation   @FHandle
